+++ digital-ocean/fedora-coreos/kubernetes/network.tf
@@ -39,14 +39,7 @@ resource "digitalocean_firewall" "rules" {
     source_tags = [digitalocean_tag.controllers.name, digitalocean_tag.workers.name]
   }
 
-  # IANA vxlan (flannel)
-  inbound_rule {
-    protocol    = "udp"
-    port_range  = "4789"
-    source_tags = [digitalocean_tag.controllers.name, digitalocean_tag.workers.name]
-  }
-
-  # Linux vxlan (Cilium)
+  # vxlan
   inbound_rule {
     protocol    = "udp"
     port_range  = "8472"