+++ google-cloud/fedora-coreos/kubernetes/network.tf
@@ -75,16 +75,16 @@ resource "google_compute_firewall" "allow-apiserver" {
   target_tags   = ["${var.cluster_name}-controller"]
 }
 
-# flannel VXLAN
-resource "google_compute_firewall" "internal-vxlan" {
+# flannel
+resource "google_compute_firewall" "internal-flannel" {
   count = var.networking == "flannel" ? 1 : 0
 
-  name    = "${var.cluster_name}-internal-vxlan"
+  name    = "${var.cluster_name}-flannel"
   network = google_compute_network.network.name
 
   allow {
     protocol = "udp"
-    ports    = [4789]
+    ports    = [8472]
   }
 
   source_tags = ["${var.cluster_name}-controller", "${var.cluster_name}-worker"]